core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "bookshelf-core"
version = "0.1.0"
description = "Bookshelf API — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "python-dotenv>=1.2.1",
    "supabase>=2.28.0",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
